[keyserver][lib] Add Neynar config for keyserver/web/landing
Summary:
It isn't needed on `landing` yet, but it gets included automatically so I figured that there's no downside.

For `keyserver` it will be accessed via `getCommConfig`. On `web` and `landing` it will be access via `process.env.COMM_NEYNAR_KEY`.

It will be configured either by setting `keyserver/secrets/neynar.json`, or by the env var `COMM_JSONCONFIG_secrets_neynar`.

This approach follows what we're done for the Alchemy and WalletConnect API keys.

This addresses [ENG-7424](https://linear.app/comm/issue/ENG-7424/integrate-neynar-api-on-web-to-be-able-to-resolve-farcaster-usernames).

Test Plan:
I tested this in combination with later diffs on both keyserver and web:

1. I specified `keyservers/secrets/neynar.json`
1. For keyserver, I tested that Farcaster usernames worked in notifs
2. For web, I tested that Farcaster usernames were resolved in robotext

Reviewers: ginsu

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D11874