[lib] Introduce getFCNames
Summary:
This was modeled after `getENSNames`, and is an async function that will be used on the keyserver for notifs.

In the React context, we'll use a different utility that will be introduced in the next diff that allows the FC names to stream in, rather than blocking on every single one of them being fetched.

Depends on D11876

Test Plan:
I tested this in combination with later diffs on both keyserver and web:

1. For keyserver, I tested that Farcaster usernames worked in notifs
2. For web, I tested that Farcaster usernames were resolved in robotext

Reviewers: ginsu

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D11877