[native] Factor out handleReduxMigrationFailure
Summary:
This is the `native` equivalent of D11884.

In a later diff I want to call this from another file, and need to extract it in order to avoid an import cycle.

Pulling `persistBlacklist` in as well to avoid an import cycle.

Depends on D11885

Test Plan: Flow, and the testing done elsewhere in the stack

Reviewers: ginsu

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D11886