[native] Don't send reserved ETH addresses to registration flow
Summary:
This fixes an issue I identified in [ENG-8044](https://linear.app/comm/issue/ENG-8044/cant-log-in-as-commeth-to-staging-identity), which is that we send ETH accounts that have been reserved with the primary keyserver to the registration flow. These accounts should not need to go through the registration flow; rather, they should be able to transparently login via the reserved usernames workflow.

Note that after this diff, the login will still fail. [ENG-4033](https://linear.app/comm/issue/ENG-4033/update-clients-to-try-new-registration-after-failed-login-due-to) tracks updating the logic here (and in `ExistingEthereumAccount`) to try to register the account via the reserved usernames workflow (rather than a simple login).

I considered factoring this logic out, since both `ConnectEthereum` and `FullscreenSIWEPanel` need to check this. But in ENG-4033 we'll probably need to refactor that check anyways, since we'll want to behave differently in the case of a reserved username versus a fully-registered account.

Would normally put @varun on this, but as he's out I figured @inka would be a good reviewer.

Test Plan:
Try logging in as comm.eth on staging. Confirm that it simply fails now, instead of redirecting the user to the registration flow

Reviewers: inka

Subscribers: