[keyserver] introduce a table for communities
Summary: This new table will be used to store information about a community. For the scope of the project, I intentially made the schema for the table very limited + only focused on tagging a farcaster channel, and figured we could always alter the table in the future to include other things like invite links when I have more free cycles

Test Plan: Ran migration and confirmed the table was created. Also created a new db and confirmed that this also creates the communites table.

Reviewers:

Subscribers: