[keyserver] modify thread creator + deleter endpoints to update communities table
Summary: With our new communities table on the keyserver we want to make sure that if a root level thread is created or deleted (a communitiy) that we also create + delete the respective row in the communities table.

Test Plan: Created a thread through the community creation user flow + confirmed that a new row was inserted in both the `threads` and `communities` tables. Deleted that root thread (communitiy) and confirmed that the correct row was removed in both the `threads` table and the `communities` table

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D11904