[Terraform] Give backup service AWS IAM role necessary to query and update backup service logs table
Summary: Staging and production backup service doesn't have necessary permissions to update backup service logs. This is the reason why log upload and download on iOS is broken. This differential adds necessary permission so that the logs functionality is fixed on iOS.

Test Plan:
Prior to the test plan I had to delete and recreate backup-service-logs table since its schema was out of sync with TF and service source code (`logID` was still considered a string)
1. Run terraform apply.
2. Upload backup from iOS device.
3. Start typing draft after uploading to trigger log upload.
4. Log out.
5. Log in.
6. Restore.
7. Ensure that draft is restored.

Reviewers: bartek

Subscribers: