[lib] add reserved username workflow to identity login actions
Summary: in our identity login actions, we should check the exception message and if it tells us that the username/wallet address is reserved, we should attempt reserved registration (if possible, i.e. if on native)

Depends on D11911

Test Plan: called the useLogIn hooks with a reserved username and reserved wallet address and confirming that i was registered on identity. this will be tested further in next diff which will use the login hook from the native handler component responsible for background authing users who are missing a CSAT

Reviewers: ashoat, tomek

Subscribers: