[native] implement onPressTagChannel for TagUnfollowedFarcasterChannel
Summary:
This diff implements the logic for `onPressTagChannel` in `TagUnfollowedFarcasterChannel`. The important thing about this method is that before dispatching our action promise, we want to validate the input through `fetchFarcasterChannelByName`

Depends on D11857

Test Plan: Confirmed that `fetchFarcasterChannelByName` correctly validates the input and `createOrUpdateFarcasterChannelTag` action is dispatched when the input is valid. Otherwise confirmed that I was logging the error message

Reviewers: inka, ashoat

Reviewed By: ashoat

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D11866