[native] try to log in if native user is missing CSAT
Summary: if a native password user is missing a CSAT (and usingCommServicesAccessToken is true), we should try to log in the identity and get a new CSAT

Test Plan: using this [patch](https://gist.github.com/vdhanan/e1768cdec753add5d64a2d4027d8f177), i cleared the access token on native and confirmed that i was logged in to identity again and received a new access token. i confirmed that if identity is unavailable, i get logged out and the alert displays explaining why i was logged out.

i also updated staging to use my local keyserver's public key for signing verification and then registered a user the old way to add the user to the reserved usernames table on staging.

then i flipped usingCommServicesAccessToken to true and confirmed that my redux store was updated with an access token and the reserved user was moved to the users table in staging DDB.

Reviewers: ashoat, tomek

Subscribers: