[native] try to log in if native user is missing CSAT
Summary: if a native password user is missing a CSAT (and usingCommServicesAccessToken is true), we should try to log in the identity and get a new CSAT

Test Plan:
using this [patch](https://gist.github.com/vdhanan/e1768cdec753add5d64a2d4027d8f177), i cleared the access token on native and confirmed that i was logged in to identity again and received a new access token. i confirmed that if identity is unavailable, i get logged out and the alert displays explaining why i was logged out.

also confirmed that my avatar doesn't get removed, since we perform keyserver auth as part of the `usePasswordLogIn` hook

Reviewers: ashoat, tomek

Reviewed By: ashoat

Differential Revision: https://phab.comm.dev/D12128