[keyserver] Stop creating two cookies on siwe registration
Summary:
issue: [[https://linear.app/comm/issue/ENG-8052/refactor-siweauthresponder-to-stop-creating-two-cookies-on | ENG-8052]]
We want to match how `keyserverAuthResponder` was refactored. We want to stop creating two cookies in case of registration. The cookie is created later by `processSuccessfulLogin`.

Test Plan:
Tested that it is possible to register a user with siwe ([[https://linear.app/comm/issue/ENG-8044/cant-log-in-as-commeth-to-staging-identity | ENG-8044]] still occurs, but the user is added to the table and receives a user cookie).
Tested that it is possible to login with siwe

Reviewers: ashoat, tomek

Reviewed By: ashoat

Differential Revision: https://phab.comm.dev/D11919