[native] Create a function that processes all the ops
Summary:
Create a single vector where all the ops are stored and process them in a single transaction.
This function isn't ideal, but I have discovered a lot of ways of how it can't be implemented. I checked a lot of ideas from Internet, but I haven't found anything that compiles (mainly due to a fact that we're storing a vector of unique pointers - but avoiding the pointers also requires a lot of work).

https://linear.app/comm/issue/ENG-6004/wrap-all-store-operations-from-one-action-in-transaction

Depends on D11902

Test Plan: Modified redux-utils and used this function to process the drafts operations. Then created a draft, closed and reopened the app and checked if the modified draft was present.

Reviewers: kamil, marcin, varun

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D11921