[native] Don't call handleReduxMigrationFailure for legacy migrations
Summary:
@kamil [explains this well](https://phab.comm.dev/D11887#340728) in D11887. Most calls to `unshimClientDB` (all except for the new one introduced in that diff) get passed an older version of the Redux state, where `cookie` is a top-level field instead of being inside the `KeyserverStore`.

This diff changes the default behavior of `unshimClientDB` so that it clears that top-level field, instead of calling `handleReduxMigrationFailure`, which assumes that the `cookie` field has been migrated to inside the `KeyserverStore`.

Passing this as a third param might not be the most ideal way of doing this, but it's consistent with how we do it [elsewhere](https://github.com/CommE2E/comm/blob/2b5b72a18455c113d45bbec52f23bf6cad054f39/native/redux/deprecated-client-db-utils.js#L38).

Test Plan:
1. Flow
2. Careful review of the code
3. Observation that the same approach is already used elsewhere in the codebase

I considered trying to run the actual migrations, but they are very old and this would be rather time-consuming and likely to have issues. The change here is simple enough that I think the above three things, plus code review should be sufficient

Reviewers: kamil, tomek

Reviewed By: tomek

Subscribers: kamil

Differential Revision: https://phab.comm.dev/D11928