[lib] Add UserInfosHandler
Summary:
issue: [[https://linear.app/comm/issue/ENG-7743/create-a-handler-that-will-fetch-missing-usernames | ENG-7743]]
Add a scaffolding of a handler component that will be used to fetch usernames and other user infos for entriens in the user store that only have an id.
It is unusual to be adding such empty diffs, but @kamil will need this handler in his work, so I am creating it now to avoid conflicts

Test Plan: Rendered this handler in `App`, manually added user infos `{id: '1000'}, {id: '5000'}`. Logged `userInfosWithMissingUsernames` in the `useEffect` and checked that the two userinfos were present, and inly them.

Reviewers: kamil, tomek

Subscribers: ashoat, kamil

Differential Revision: https://phab.comm.dev/D11932