[native] Set the DB version after registration
Summary:
When a new account is created, we have to store the version in the DB so that it is included in a backup.
We don't need to set the version when a user logs in, because we can assume that eventually, we will restore the data from a backup - and that includes the version.

https://linear.app/comm/issue/ENG-7006/back-up-redux-version

Depends on D11937

Test Plan: Register a new user and check if the version is present in Redux and in the DB.

Reviewers: kamil, marcin, inka

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D11942