[web] Check backup version when restoring
Summary:
If a backup is from newer version that the client, we can't restore it and the client needs to be updated in order to restore the backup.

{F1799090}

https://linear.app/comm/issue/ENG-7008/detect-that-the-compaction-is-from-a-newer-version-than-the-app-and

Depends on D12013

Test Plan: Created a backup on native and tried to restore it on web, where the native version was greater that the web version. Verified that an exception was thrown during the restoration.

Reviewers: kamil, marcin, inka

Reviewed By: kamil, inka

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D11947