[identity] modify get user identity rpc to take user ids and return identity map with farcaster id
Summary:
This diff modifies our protos for the get user identity rpc to take in a list of user ids and return their corresponding identities.

Additionally, for our farcaster work, the Identity message will now also include a farcaster id field

Depends on D11951

Test Plan:
Deployed on staging and called the rpcs with a list of user ids, including a default username user, an ethereum user, a default user with farcaster id, and a non-existent user.

The rpc returned all the proper `Identity` info and didn't include the non-existent userID

Reviewers: bartek, kamil, ashoat

Reviewed By: bartek, ashoat

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D11952