[native] Swap order of ConnectEthereum and ConnectFarcaster in registration flow
Summary: We'll be requiring a Farcaster account for the initial beta, so it makes sense to ask about it first.

Depends on D11972

Test Plan: Tested opening the registration flow from the main screen and from `FullscreenSIWEPanel`. Make sure `CoolOrNerdModeSelection` and `KeyserverSelection` don't appear. Make sure that I am able to successfully register

Reviewers: ginsu, inka

Subscribers: