[keyserver] Roll back D11838
Summary: This solved [ENG-8072](https://linear.app/comm/issue/ENG-8072/requests-to-ashoats-keyserver-are-timing-out-in-production) in production. It appears that performance was significantly affected by converting this column's type. We'll need to find another approach here.

Test Plan: I deployed this to production and confirmed that it solved [ENG-8072](https://linear.app/comm/issue/ENG-8072/requests-to-ashoats-keyserver-are-timing-out-in-production)

Reviewers: tomek, ginsu, will

Reviewed By: ginsu

Differential Revision: https://phab.comm.dev/D11976