[native] Create a function returning the DB version
Summary:
We're storing the version in synced metadata table. We don't want to fetch all the data, which means that we have to introduce a dedicated function for this.

https://linear.app/comm/issue/ENG-7008/detect-that-the-compaction-is-from-a-newer-version-than-the-app-and

Depends on D11942

Test Plan: Tested with the next diff - verified that the version returned by this function matched what was in the backup.

Reviewers: kamil, marcin, varun

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D12012