[lib][native] Add nonceTimestamp to SIWEResult
Summary:
`SIWEResult` is a type only used on `native`. It's what `SIWEPanel` returns after a successful SIWE.

For [ENG-7665](https://linear.app/comm/issue/ENG-7665/dont-display-use-connected-ethereum-wallet-if-nonce-was-generated-more), we want to know when the nonce was generated in order to determine whether to allow the user to use the cached SIWE result.

Would normally put @varun on the review here, but going to ask @inka instead because he is out.

Test Plan:
I tested (or will test) this diff stack as follows:

1. Be in a multi-keyserver environment, testing SIWE with iOS simulator
2. Do a SIWE and then wait 2 minutes to let the nonce expire in the following screens:
  - `FullscreenSIWEPanel` for an account that doesn't exist yet when the new registration flow is disabled
  - `FullscreenSIWEPanel` for an account that doesn't exist yet when the new registration flow is enabled
  - `FullscreenSIWEPanel` for an account that does exist
  - New registration flow for an account that doesn't exist yet (`RegistrationTerms`)
  - New registration flow for an account that does exist (`ExistingEthereumAccount`)
3. Make sure there are no duplicate `Alert`s, that in all cases an `Alert` is shown, and that the "back" action activates when the user confirms the `Alert`

Reviewers: inka

Subscribers: