[native] Display meaningful alert if ExistingEthereumAccount login fails due to expired nonce
Summary: This addresses part of [ENG-7666](https://linear.app/comm/issue/ENG-7666/display-meaningful-alert-if-wallet-registration-fails-due-to-expired), specifically for `ExistingEthereumAccount`. Doing this separately since it's the "easy part" and introduces some complicated navigation `getParent()` type parameterization that will be used in the next diff.

Would normally put @varun on the review here, but going to ask @inka instead because he is out.

Depends on D12008

Test Plan:
I tested (or will test) this diff stack as follows:

1. Be in a multi-keyserver environment, testing SIWE with iOS simulator
2. Do a SIWE and then wait 2 minutes to let the nonce expire in the following screens:
  - `FullscreenSIWEPanel` for an account that doesn't exist yet when the new registration flow is disabled
  - `FullscreenSIWEPanel` for an account that doesn't exist yet when the new registration flow is enabled
  - `FullscreenSIWEPanel` for an account that does exist
  - New registration flow for an account that doesn't exist yet (`RegistrationTerms`)
  - New registration flow for an account that does exist (`ExistingEthereumAccount`)
3. Make sure there are no duplicate `Alert`s, that in all cases an `Alert` is shown, and that the "back" action activates when the user confirms the `Alert`

Reviewers: inka