[native] Run backup migrations
Summary: Run the Redux migrations using the default state and a restored SQLite. We run these only because they are modifying the SQLIte and we don't care about the Redux state that was transformed - we're using the default state and throw away the result.

https://linear.app/comm/issue/ENG-7009/run-redux-migrations-after-the-backup-is-restored

Depends on D12015

Test Plan: Create a backup, kill the app, add a new migration, open the app, restore the backup and check if the migration was run - by calling `commCoreModule.getStoredVersion()` before and after the migrations.

Reviewers: kamil, marcin, varun