[lb] introduce fetchFarcasterChannelInfo to neynar client class
Summary: In the communities table we store the farcaster channel id. We need to use that id to fetch the channel info so we can display the channel name to the user. Neynar has an endpoint for this and this diff introduces the function that calls this neynar endpoint and fetches the channel info

Test Plan: Passed in several channel ids to this method and confirmed that I was getting the correct response back

Reviewers:

Subscribers: