[native] display existing farcaster channel tag in TagFarcasterChannel screen
Summary:
When the user goes to the `TagFarcasterChannel` screen they should be able to see if the community already has a farcaster channel already associated with it. This diff uses the `farcasterChannelID` stored from the keyserver to fetch the farcaster channel info for the given `farcasterChannelID`, and uses that info to display the farcaster channel name if their is an existing tag. Otherwise it will let the user know that there are currently no tagged farcaster channels tagged to that community.

The mechanism for getting the farcaster channel info that I came up with is a bit naive and not very performant since we are fetching the channel info pretty often; however, I had to make sure I didn't blow up the scope of this project, and I felt that the tradeoff to ship this faster over slightly improved performance was necessary. If this implementation is okay, I will also create a linear task before landing to track making this mechanism more performant.

Depends on D12026

Test Plan:
Please see demo video below (Remove button will be introduced in a separate diff)

{F1771871}

Reviewers: tomek, inka, ashoat

Reviewed By: ashoat

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D12041