[native] display existing farcaster channel tag in TagFarcasterChannel screen
Summary: When the user goes to the `TagFarcasterChannel` screen they should be able to see if the community already has a farcaster channel already associated with it. This diff uses the `farcasterChannelID` stored from the keyserver to fetch the farcaster channel info for the given `farcasterChannelID`, and uses that info to display the farcaster channel name if their is an existing tag. Otherwise it will let the user know that there are currenlty no tagged farcaster channels tagged to that community

Test Plan: Please see demo video below (Remove button will be introduced in a seperate diff)

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D12041