[keyserver][nix] remove dbType from db-config.js
Summary:
This removes the usage of dbType from keyserver's db_config.js file. This is no longer necessary and we plan on upgrading from mariadb10.8 to 10.11.

There is a warning that we're supplying dbType as a mariadb configuration option that doesn't exist (as db-config.js now contains a field we don't use), but looks like devs can just manually remove dbType to remove this warning after this is landed

Test Plan: Ran keyserver after removing dbType and ensured it was working with iOS simulator

Reviewers: ashoat, O1 Nix

Reviewed By: ashoat

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D12021