[nix] upgrade mariadb from 10.8 to 10.11
Summary:
The modification of the flake.nix file is just removing the mariadb v.10.8 overlay, automatically bumping us to v.10.11.

The mariadb-up script is also modified to check whether or not an upgrade is necessary. If required, it will run `mariadb-upgrade`.

For a dev currently on 10.8, after this is landed, the dev will
1. pull the commit
2. exit their terminals and any `nix develop` sessions
3. pkill mariadb` if there is a running server process (`ps aux | grep mariadb`)
4. nix develop

On this nix develop, it should identify an upgrade is required and print out upgrade progress

After this step, everything should be upgraded

Depends on D12021

Test Plan:
- Set up dev environment with a mariadb 10.8 versioned table
- Launched an iOS simulator and exchanged messages between test users on a test community
- Ran through the above upgrade steps
- On a new iOS simulator, ensured that the messages were persisted from before and that new communities and users could be registered and exchange messages

Reviewers: ashoat, O1 Nix

Reviewed By: ashoat

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D12022