[nix] upgrade mariadb from 10.8 to 10.11
Summary:
The modification of the flake.nix file is just removing the mariadb v.10.8 overlay, automatically bumping us to v.10.11.

The mariadb-up script is also modified to check whether or not an upgrade is necessary. If required, it will run `mariadb-upgrade`.

For a dev currently on 10.8, after this is landed, the dev will
1. pull the commit
2. exit their terminals and any `nix develop` sessions
3. pkill mariadb` if there is a running server process (`ps aux | grep mariadb`)
4. nix develop

On this nix develop, it should identify an upgrade is required and print out upgrade progress

After this step, everything should be upgraded

If the dev has important information on their local keyserver that they want to ensure is backed up, they can run through Inka's backup documentation:
https://www.notion.so/commapp/Backing-up-a-db-21d2a4a896f7496da3e32f19a137aaf5
https://www.notion.so/commapp/Restoring-a-db-8eff886db5f742a0ac5a3ccda0722c70

Depends on D12021

Test Plan:
- Set up dev environment with a mariadb 10.8 versioned table
- Launched an iOS simulator and exchanged messages between test users on a test community
- Ran through the above upgrade steps
- On a new iOS simulator, ensured that the messages were persisted from before and that new communities and users could be registered and exchange messages

Reviewers:
ashoat

Subscribers: