[keyserver] Script to print largest Farcaster communities
Summary:
Rahul needed this for some of his work so I whipped up a quick script. Wanted to commit it to the codebase in case we need to re-run it later to get an updated list.

Details in [OPS-283](https://linear.app/comm/issue/OPS-283/write-script-to-determine-largest-farcaster-channels-by-follower-count).

Test Plan: I ran the script. Results are in the Linear task

Reviewers: ginsu

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D12020