[native] introduce new updateRolesAndPermissions migration util
Summary: The last step for introducing the communties table in the keyserver + tagging farcaster channel project is we need to introduce a new permission to manage the farcaster channel tags. Part of this task will be to perform a redux migrations so that native clients add this permisssion manage farcaster tag permission. However, as part of the work to minimally encode the permissions, `updateRolesAndPermissions` became `legacyUpdateRolesAndPermissions` since it

Test Plan: Ran the migration in the very next diff and logged out `threadStoreInfos` at the very begining of the method and then logged out `updatedThreadStoreInfos` at the very end of the method. Put both logs into a json compare tool and the diff between the two json objects looked reasonable.

Reviewers:

Subscribers: