[lib] Bind output validators to endpoints on lib
Summary:
issue: [[https://linear.app/comm/issue/ENG-8080/create-a-mapping-between-endpoints-and-output-validators-in-lib | ENG-8080]]
As I explained in the issue, I wanted to simply use those objects instead of objects in lib/types/endpoints.js. But this causes very complicated import cycles. I know how to solve them, but it would requre refactoring a lot of files, hundreds possibly.
So instead I am creating a separate file with this mapping endpoint-output validator. The dev will not be able to forget about adding the validator here, because that would cause flow issues in the next diff.

Test Plan: Tested with next diffs - checked a couple of endpoints, and verified that no errors are thrown if outputs are validated with validators specified by `endpointValidators` for those endpoints.

Reviewers: tomek, kamil, bartek

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D12030