[lib] Add a way of marking migrations as schema-changing
Summary: When running the migrations check if any of them is marked as schema-changing.

https://linear.app/comm/issue/ENG-7493/add-a-way-of-marking-migrations-as-schema-changing

Depends on D12048

Test Plan: Add a log checking if `schemaChanged` is computed correctly. Add a migration without `changesSchema` and check if the value is false. Then add a migration that `changesSchema` and check if the value is true.

Reviewers: kamil, marcin, varun