Modify Rust restore code to avoid fetching latest backup id twice
Summary: This differentila modifies the C++ and Rust backup code so that it is possible to restore without querying the backup service for backup id if it is already provided. For SIWE users we will know backup id before we want to restore since we need backup message (that comes with backup id) to prepare backup secret.

Test Plan: For now check that restoration for non-password users work. Next diff in the stack will test SIWE case.

Reviewers: kamil, bartek

Subscribers: