[lib] Create a compaction after schema-changing migration
Summary: Check if any of migrations changes schema and create a compaction if that happened.

This diff introduces a new flag that will be set to true when all the primary devices will be backed up (the switch from dev menu is removed).

https://linear.app/comm/issue/ENG-7014/create-new-compaction-after-sqlite-migration

Depends on D12049

Test Plan: Enable backup, set the flag to true, run a migration that changes schema and check in the backup service if the backup was created.

Reviewers: kamil, marcin, varun