[native] Consume `useAvailableThreadMemberActions` in `native/ThreadSettingsMember`
Summary:
Similar to D11913, but for `ThreadSettingsMember`.

This was last usage of `getAvailbleThreadMemberAction`, so we also go ahead and remove it.

---

Depends on D11913

Test Plan:
`flow` + close reading for now

Logged value of `getAvailbleThreadMemberAction` and `useAvailableThreadMemberActions` and ensured value was the same. Assuming the logic in `ThreadSettingsMember` that works with value is and remains correct.

Reviewers: ashoat, ginsu, tomek, varun, will

Reviewed By: ashoat

Differential Revision: https://phab.comm.dev/D11914